move animation frame type
